Finished Moderator

+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/AlbumDao.java
@@ -12,7 +12,7 @@ public interface AlbumDao {
     List<Album> findAll();
     List<Album> findByTitleContaining(String sub);
 
-    int save(Album album);
+    long save(Album album);
 
     int update(Album album);
